Enhanced controller
Added:
LIDAR support:
by using an LD06 lidar one can create a 'plane' to check for intrusions rendering the space unsafe. This acts a s a security check in case the telescope in the way of the roof.
added UI elements to index and setup to enable this feature including a canvas to visualize the points

Dome security checks:
- check if the mount is parked by querying the ONSTEP mount
- checks if the LIDAR reports safe
- strict or lose security check: in strict mode all checks must be safe in lose mode only one check can pass
added checkboxes to the setup webpage to enable the features

ESP32-S3 board support.
for switches certain ESP32 boards have higher pin numbers

Translated parts of the UI to English

changed the behavior for Aleko gate openers that only activate on falling signal